[EN] connect-to-broker without an atom
There was a discussion on github in the PR and we decided to try
different approaches. There was the atom approach, a lazy seq approach,
and this approach, which moves the try/catch immediately around the
connectiona attempt. The catch logs the error and returns nil.

If we have a connection, we return it, otherwise we recur.