default to more typical queue options in wire-up/external-service
These are the options we use everywhere, and seem to make more sense to me even in the context of this open-source, general-purpose library. But others feel differently, there are other ways to improve upon this. Currently the workaround is to use the higher-arity version everywhere. But as it is, I'm not sure the lower-arity version would be useful to anyone (or at least, it isn't to us).