Recovery of long lived server-named queues
Server-named queues set up by kehaar for use with responders need to
be tracked so that kehaar can set the correct reply-to queue for
requests when those queues get renamed during a Rabbit recovery from a
failure state.

This keeps track of a mapping from an external service queue name to
the name of the server-named queue for responses. State is kept in an
atom holding a bidirectional map, and convenience functions are
provided for interacting with it.

When Rabbit renames a server-nemed queue, a callback updates the state
if necessary.

Short-lived server-named queues are not tracked. You better believe
that a service will time out waiting for an answer before Rabbit can
recover. It is not fast.