Allow "handling" threads to be released when not in use
Replace old `thread-handler` code that pre-allocated a pool of threads,
each trying to take work from a core.async channel with a system that
spawns threads as work is available, using a semaphore to limit the
number of live threads at a given time.

The threads spawned now will be deallocated at the end of the work
they do[1], making it easier to find threads actually doing stuff in a
Java profiler.

[1]: Well, core.async is a little smarter than just deallocating a
thread when it's finished. it will be held on to for a time and reused
if new work comes around quickly. After some amount of time, if no new
work shows up, it'll be killed.