Add blank / nil config param helper fn
Ran into a need for something like this when starting to use CloudAMQP (which needs non-default values for some of these params where our legacy clusters do not). Figured it was better in here than repeated all over our components (which already happened with an earlier, less generic version of this fn).