Remove the not-yet-requested tests of realize-fn
We're testing that the ns is required, but as soon as the first test
runs, the ns _is_ required, so those two tests weren't doing what they
claimed.

And now that this test is back to using vanilla realize-symbol-or-self
(and doesn't need to be a function), `kehaar.not-yet-required/foo` can
be a plain `def` again as well.