Do not recover streaming response channels
The RabbitMQ Java client does channel and topology recovery when it
detects connection issues. This includes declaring any queue that was
declared by this connection, but that was not deleted by the connection.
Since bespoke streaming response queues are created by the responder,
but deleted by the consumer, the responder has no way of knowing that
the queue was deleted. This means that when the responder handles
automatic recovery, it re-declares every bespoke response queue it has
ever created.